Auto stash before checking out "origin/development"
