+++ src/main/java/Services/UserPreferenceService/UserPreferenceService.java
@@ -23,176 +23,26 @@
  */
 package Services.UserPreferenceService;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.InvalidPathException;
-import java.text.MessageFormat;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import com.google.gson.Gson;
-
 /**
  * The {@linkplain UserPreferenceService} handles user preferences
  */
-public class UserPreferenceService implements IUserPreferenceService
+public class UserPreferenceService extends UserPreferenceBaseService<UserPreference> implements IUserPreferenceService
 {
     /**
-     * The current class logger
-     */
-    private final Logger logger = LogManager.getLogger();
-    
-    /**
-     * The name of folder where user preference is storage.
-     */
-    public static final String UserPreferenceDirectoryName = "UserPreferences";
-
-    /**
-     * The setting file extension
-     */
-    public static final String SettingFileExtension = ".settings.json";
-    
-    /**
-     * Backing field for {@linkplain GetUserPreference()}
-     */
-    private UserPreference userPreference;
-    
-    /**
-     * Gets the {@linkplain IUserPreference} the {@linkplain UserPreferenceService} currently write and read from/to
-     * 
-     * @return a {@linkplain IUserPreference}
+     * Gets the {@linkplain Class} of {@linkplain #TUserPreference}
      */
     @Override
-    public UserPreference GetUserPreference()
+    protected Class<UserPreference> GetUserPreferenceType()
     {
-        if(this.userPreference == null)
-        {
-            this.Read();
-        }
-        
-        return this.userPreference;
+        return UserPreference.class;
     }
-    
+
     /**
-     * The {@linkplain File} that holds the user preferences
-     */
-    private File userPreferenceFile = this.GetSettingFile();
-    
-    /**
-     * Reads the {@linkplain IUserPreference} from the file
-     * 
-     * @return the {@linkplain UserPreference}
+     * Gets the user preference file name
      */
     @Override
-    public UserPreference Read()
+    protected String GetFileName()
     {
-        if (!this.TryCreateDirectoryOrFile(this.userPreferenceFile))
-        {
-            throw new InvalidPathException(this.userPreferenceFile.getAbsolutePath(), "Cannot make use of the user preference service. Check the log for more detail"); 
-        }
-
-        try (FileInputStream reader = new FileInputStream(this.userPreferenceFile))
-        {
-            byte[] data = new byte[(int) this.userPreferenceFile.length()];
-            int read = reader.read(data);
-            
-            this.logger.info(String.format("%s bytes read from the UserPreference file ", read));
-            
-            String dataString = new String(data, StandardCharsets.UTF_8);
-            
-            this.userPreference = new Gson().fromJson(dataString, UserPreference.class);
-            
-            if (this.userPreference == null)
-            {
-                this.userPreference = new UserPreference();
-            }
-            
-            return this.userPreference;
-        }
-        catch (IOException exception)
-        {
-            this.logger.error(MessageFormat.format("Could not read the user setting file {0} because {1}", 
-                    this.userPreferenceFile.getAbsolutePath(),
-                    exception));
-            
-            return null;
-        }
+        return "";
     }
-    
-    /**
-     * Saves the {@linkplain userPreference} to disk
-     */
-    @Override
-    public void Save()
-    {
-        if (!this.TryCreateDirectoryOrFile(this.userPreferenceFile))
-        {
-            throw new InvalidPathException(this.userPreferenceFile.getAbsolutePath(), "Cannot make use of the user preference service. Check the log for more detail"); 
-        }
-        
-        try (FileWriter writer = new FileWriter(this.userPreferenceFile.getAbsolutePath(), false))
-        {
-            writer.write(new Gson().toJson(this.userPreference)); 
-        } 
-        catch (IOException exception)
-        {
-            this.logger.error(MessageFormat.format("Could not write the user setting file {0} because {1}", 
-                    this.userPreferenceFile.getAbsolutePath(),
-                    exception));
-        }
-    }
-        
-    /**
-     * Gets the setting file
-     * 
-     * @return a {@linkplain File}
-     */
-    private File GetSettingFile()
-    {
-        File file = new File(System.getProperty("user.home"));
-        file = new File(file, ".rheagroup");
-        this.TryCreateDirectoryOrFile(file);
-        file = new File(file, "DEHAdapterSettingFile");
-        this.TryCreateDirectoryOrFile(file);
-        file = new File(file, UserPreferenceDirectoryName);
-        this.TryCreateDirectoryOrFile(file);
-        file = new File(file, UserPreference.class.getSimpleName() + SettingFileExtension);
-        this.TryCreateDirectoryOrFile(file);
-        return file;
-    }
-
-    /**
-     * Create the directory if it does not exist yet
-     * 
-     * @param fileOrDirectory the {@linkplain File} containing the path
-     * @return a value indicating whether all checks on the {@linkplain file} are ok 
-     */
-    private boolean TryCreateDirectoryOrFile(File fileOrDirectory)
-    {
-        try 
-        {                
-            if(fileOrDirectory.getName().endsWith(SettingFileExtension) && !fileOrDirectory.exists())
-            {
-                return fileOrDirectory.createNewFile();
-            }
-            
-            if(!fileOrDirectory.exists() && fileOrDirectory.mkdir())
-            {
-                return fileOrDirectory.createNewFile();
-            }
-            
-            return fileOrDirectory.exists();
-        }
-        catch (IOException exception)
-        {
-            this.logger.error(MessageFormat.format("Could not create specified directory or file: {0}", fileOrDirectory.getAbsolutePath()));
-            this.logger.catching(exception);
-            return false;
-        }
-    }
-
 }
